begining refactoring to externalize json commands
